requirements.txt
aniso8601==9.0.1
attrs==21.4.0
blinker==1.5
boto3==1.21.45
botocore==1.24.45
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
flask-restx==0.5.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
jsonschema==4.4.0
MarkupSafe==2.1.1
mongoengine==0.24.1
Pillow==9.1.0
pyhumps==3.7.2
PyJWT==2.3.0
pymongo==4.1.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
